Fix binary expression rendering when group modifiers are involved
Parentheses that follow a `group_left` or `group_right` can be ambiguous
when they surround the right side of the binary expression and when no
argument is passed to the group modifier.

This fix explicitely renders empty parentheses for group modifiers that
have no label list.

Same as edf5d3ce8ddc577cfd949a7ec67b008a3f8d0f54, which was for Python.
This commit is the same fix for Go and Typescript.